mcf-ads-engine/pyproject.toml
[project]
name = "mcf-ads-engine"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "google-ads>=24.0.0",
    "google-auth-oauthlib>=1.2",
    "anthropic>=0.30.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "resend>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "numpy>=1.26",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
